# doc/api.yaml
openapi: 3.0.3
info:
  title:  WASA Photo
  description: |
    WASAPhoto REST APIs
  version: "1.0.0"
...
paths:
...
  /profiles/{user_name}/photos:
    parameters:
      - name: user_name
        in: path
        description: Describe the user name
        required: true
        schema:
          type: string
          example: Maria 
          minLength: 3
          maxLength: 16
          pattern: '^.*?$'
    post:
      tags:
        - photo
      operationId: uploadPhoto
      summary: upload a photo
      description: This method allows the user to upload a photo
      requestBody:
        required: true
        description: The image we want to upload
        content:
          image/png:
            schema:
              type: string
              format: binary
              example: "0000100100100111000101001110111000101010001110101"
              minLength: 20
              maxLength: 10000000
              pattern: '^.*?$'
              description: The photo the user want to upload
      responses:
        "201":
          description: The photo has been uploaded and its information is being returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/photo"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    get:
      tags:
        - photo
      operationId: getMyStream
      summary: get my stream of photo
      description: The method allows the user to know his stream (in reverse chronological order)
      responses:
        "200":
          description: Return the photo stream
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/photo"
                minItems: 0
                maxItems: 50
                description: The photo stream
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT 
  responses:
    BadRequest:
      description: The request was not complaint with the documentation
    InternalServerError:
      description: The server encountered an internal error
    Unauthorized:
      description: You do not have permission to do this operation
...
  schemas:
...
    photo:
      title: photo
      description: description of a photo
      type: object
      properties:
        photo_id:
          type: string
          example: d732fhh
          minLength: 5
          maxLength: 100
          pattern: '^.*?$'
          description: The unique photo ID
        date: 
          type: string
          example: "2022-01-01"
          minLength: 10
          maxLength: 10
          format: date
          pattern: '^.*?$'
          description: The date when the photo was uploaded
        like_number:
          type: integer
          example: 100
          description: The number of photo's likes 
        comment_number:
          type: integer
          example: 50
          description: The number of photo's comments
...
security:
  - bearerAuth: []  